Add test case for HTCondor parsing
A possible out from HTCondor-CE is a trailing pipe symbol and no scaling
factor (as it uses the pipe as a value terminator. Parser would split
this and interpret the scale factor as an empty string which would make
the parsing fail. This adds a test case to check this situation.